% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDF.R
\name{buildDF}
\alias{buildDF}
\title{Prepare nested dataset}
\usage{
buildDF(df, unit_ID, subUnit_ID)
}
\arguments{
\item{df}{dataframe}

\item{unit_ID}{character; unit column name}

\item{subUnit_ID}{character; unit column name}
}
\value{
processed df
}
\description{
Prepare nested dataset
}
