% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{getMatchCount}
\alias{getMatchCount}
\title{Percentage of successful matches vs. original units in each replacement unit group}
\usage{
getMatchCount(m.out, title = NULL)
}
\arguments{
\item{m.out}{list; selectMatch() output}
}
\value{
ggplot object
}
\description{
Percentage of successful matches vs. original units in each replacement unit group
}
