% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{getSubUnitBalance}
\alias{getSubUnitBalance}
\title{Balance of original sub-units vs. sub-units from replacement unit groups (figure)
SMD between sampled units vs. all units in population, for each covariate of interest}
\usage{
getSubUnitBalance(m.out)
}
\arguments{
\item{m.out}{list; selectMatch() output}
}
\value{
ggplot object
}
\description{
Balance of original sub-units vs. sub-units from replacement unit groups (figure)
SMD between sampled units vs. all units in population, for each covariate of interest
}
