% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{getUnitLovePlot}
\alias{getUnitLovePlot}
\title{Love plot for sampled units vs. all units in population (figure)
SMD between sampled units and all units in population}
\usage{
getUnitLovePlot(m.out)
}
\arguments{
\item{m.out}{list; selectMatch() output}
}
\value{
ggplot object
}
\description{
Love plot for sampled units vs. all units in population (figure)
SMD between sampled units and all units in population
}
