% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{matchBalance}
\alias{matchBalance}
\title{Figure: SMD between sampled units vs. all units in population, for each covariate of interest}
\usage{
matchBalance(
  m.out,
  title =
    "Standardized Mean Difference: Replacement Unit Groups (1...K) vs. Originally Selected Units"
)
}
\arguments{
\item{m.out}{list; selectMatch() output}

\item{title}{character; user-specified figure title}
}
\value{
ggplot object
}
\description{
Figure: SMD between sampled units vs. all units in population, for each covariate of interest
}
