% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{matchBalanceFig}
\alias{matchBalanceFig}
\title{Figure: SMD between sampled units vs. all units in population, for each covariate of interest}
\usage{
matchBalanceFig(m.out, title = NULL)
}
\arguments{
\item{m.out}{list; selectMatch() output}
}
\value{
ggplot object
}
\description{
Figure: SMD between sampled units vs. all units in population, for each covariate of interest
}
