% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{matchBalanceTab}
\alias{matchBalanceTab}
\title{Summary table: SMD between sampled units and population, for each covariate of interest}
\usage{
matchBalanceTab(m.out)
}
\arguments{
\item{m.out}{list; selectMatch() output}
}
\value{
ggplot object
}
\description{
Summary table: SMD between sampled units and population, for each covariate of interest
}
