% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{matchCountFig}
\alias{matchCountFig}
\title{Percentage of successful matches in each replacement unit group, 1...K}
\usage{
matchCountFig(m.out, title = "Percentage of Successful Matches per Unit Group")
}
\arguments{
\item{m.out}{list; selectMatch() output}
}
\value{
ggplot object
}
\description{
Percentage of successful matches in each replacement unit group, 1...K
}
