% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{matchCountTab}
\alias{matchCountTab}
\title{Table: Percentage of successful matches in each replacement unit group, 1...K}
\usage{
matchCountTab(m.out, title = NULL)
}
\arguments{
\item{m.out}{list; selectMatch() output}
}
\value{
dataframe
}
\description{
Table: Percentage of successful matches in each replacement unit group, 1...K
}
