% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{matchFreq}
\alias{matchFreq}
\title{Match frequency}
\usage{
matchFreq(smOut, title = "Match Frequency per Original Unit")
}
\arguments{
\item{smOut}{list; selectMatch() output}

\item{title}{character; user-specified figure title}
}
\value{
ggplot object
}
\description{
Distribution of successful matches among original units
}
