% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{summary}
\alias{summary}
\title{Build summary tables, with unit/match/sub-unit balance between initially selected units and a target population, for each covariate of interest}
\usage{
summary(m.out, balMeasure)
}
\arguments{
\item{m.out}{list; selectMatch() output}

\item{balMeasure}{numeric; balance Diagnostic (1=original unit balance,
2=match balance, 3=match success percent, 4=sub-unit balance)}
}
\value{
ggplot object
}
\description{
Build summary tables, with unit/match/sub-unit balance between initially selected units and a target population, for each covariate of interest
}
