% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMatch.R
\name{buildDF}
\alias{buildDF}
\title{Prepare dataset}
\usage{
buildDF(df, unit_ID, subUnit_ID, unit_vars, subUnit_samp_vars)
}
\arguments{
\item{df}{dataframe}

\item{unit_ID}{character; unit column name}

\item{subUnit_ID}{character; unit column name}

\item{unit_vars}{character; unit column name}

\item{subUnit_samp_vars}{character; unit column name}
}
\value{
processed df
}
\description{
Prepare dataset
}
