% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{getUnitReplacementBalance}
\alias{getUnitReplacementBalance}
\title{SMD between sampled units vs. all units in population, for each covariate of interest}
\usage{
getUnitReplacementBalance(m.out, title = NULL)
}
\arguments{
\item{m.out}{list; selectMatch() output}
}
\value{
ggplot object
}
\description{
SMD between sampled units vs. all units in population, for each covariate of interest
}
