% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{matchBalance}
\alias{matchBalance}
\title{Match balance}
\usage{
matchBalance(
  smOut,
  title = "Standardized Mean Difference:",
  subtitle = "Replacement Unit Groups (1...K) vs. Originally Selected Units"
)
}
\arguments{
\item{smOut}{list; selectMatch() output}

\item{title}{character; user-specified figure title}

\item{subtitle}{character; user-specified figure title}
}
\value{
ggplot object
}
\description{
Balance between initially sampled units and their K matches, for each covariate of interest
}
