% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{matchCount}
\alias{matchCount}
\title{Successful matches for each replacement group}
\usage{
matchCount(smtchOut, title = "Percentage of Successful Matches per Unit Group")
}
\arguments{
\item{title}{character; user-specified figure title}

\item{smOut}{list; selectMatch() output}
}
\value{
ggplot object
}
\description{
Percentage of successful matches in each unit replacement group, 1...K
}
