% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{matchFreq}
\alias{matchFreq}
\title{Successful unit match distribution among original units}
\usage{
matchFreq(m.out, title = "Match Frequency per Original Unit")
}
\arguments{
\item{m.out}{list; selectMatch() output}

\item{title}{character; user-specified figure title}
}
\value{
ggplot object
}
\description{
Successful unit match distribution among original units
}
