% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDF.R
\name{prepDF}
\alias{prepDF}
\title{Prepare nested dataset}
\usage{
prepDF(df, unitID, subunitID)
}
\arguments{
\item{df}{dataframe}

\item{unitID}{character; unit column name in original dataset}

\item{subunitID}{character; sub-unit column name in original dataset}
}
\value{
processed dataframe
}
\description{
Prepare nested dataset
}
