% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR-package.R
\docType{package}
\name{sitepickR-package}
\alias{sitepickR}
\alias{sitepickR-package}
\title{sitepickR: Two-level Sample Selection with Optimal Site Replacement}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Carries out a two-level sample selection where the possibility of an initially selected site not wanting to participate is anticipated, and the site is optimally replaced. The procedure aims to reduce the bias (and/or loss of generalizability) with respect to the target population.
}
\author{
\strong{Maintainer}: Elena Badillo-Goicoechea \email{egoicoe1@jhu.edu}

Authors:
\itemize{
  \item Robert Olsen \email{robolsen@gwu.edu}
  \item Elizabeth Stuart \email{estuart1@jhu.edu}
}

}
\keyword{internal}
