% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{subUnitBalance}
\alias{subUnitBalance}
\title{Sub-unit balance between initially selected units and all units in population, for each covariate of interest}
\usage{
subUnitBalance(
  m.out,
  title = "Subunits from Original and Replacement Unit Groups vs. Population (SMD)"
)
}
\arguments{
\item{m.out}{list; selectMatch() output}

\item{title}{character; user-specified figure title}
}
\value{
ggplot object
}
\description{
Sub-unit balance between initially selected units and all units in population, for each covariate of interest
}
