% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{unitLovePlot}
\alias{unitLovePlot}
\title{Original units balance}
\usage{
unitLovePlot(
  smOut,
  title = "Standardized Mean Difference",
  subtitle = "Initially Selected Units vs. Population"
)
}
\arguments{
\item{smOut}{list; selectMatch() output}

\item{title}{character; user-specified figure title}

\item{subtitle}{character; user-specified figure subtitle}
}
\value{
ggplot object
}
\description{
Balance between initially sampled units and all units in the population
}
