% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepickR_diagnostics.R
\name{unitLovePlot}
\alias{unitLovePlot}
\title{SMD between sampled units and all units in population}
\usage{
unitLovePlot(
  m.out,
  title = "Standardized Mean Difference: \\n Initially Selected Units vs. Population"
)
}
\arguments{
\item{m.out}{list; selectMatch() output}

\item{title}{character; user-specified figure title}
}
\value{
ggplot object
}
\description{
SMD between sampled units and all units in population
}
